// src/api/models/ApartmentImages.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SITE API
 * API сайта ПИК Брокер
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime'
import { ImageGroup, ImageGroupFromJSON, ImageGroupFromJSONTyped, ImageGroupToJSON } from './'

/**
 *
 * @export
 * @interface ApartmentImages
 */
export interface ApartmentImages {
  /**
   *
   * @type {string}
   * @memberof ApartmentImages
   */
  uuid: string
  /**
   *
   * @type {string}
   * @memberof ApartmentImages
   */
  title: string
  /**
   *
   * @type {string}
   * @memberof ApartmentImages
   */
  alt: string
  /**
   *
   * @type {boolean}
   * @memberof ApartmentImages
   */
  isMain: boolean
  /**
   *
   * @type {boolean}
   * @memberof ApartmentImages
   */
  isPlan: boolean
  /**
   *
   * @type {string}
   * @memberof ApartmentImages
   */
  fileUrl: string
  /**
   *
   * @type {ImageGroup}
   * @memberof ApartmentImages
   */
  imageGroup: ImageGroup | null
}

export function ApartmentImagesFromJSON(json: any): ApartmentImages {
  return ApartmentImagesFromJSONTyped(json, false)
}

export function ApartmentImagesFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ApartmentImages {
  if (json === undefined || json === null) {
    return json
  }
  return {
    uuid: json['uuid'],
    title: json['title'],
    alt: json['alt'],
    isMain: json['isMain'],
    isPlan: json['isPlan'],
    fileUrl: json['fileUrl'],
    imageGroup: ImageGroupFromJSON(json['imageGroup']),
  }
}

export function ApartmentImagesToJSON(value?: ApartmentImages | null): any {
  if (value === undefined) {
    return undefined
  }
  if (value === null) {
    return null
  }
  return {
    uuid: value.uuid,
    title: value.title,
    alt: value.alt,
    isMain: value.isMain,
    isPlan: value.isPlan,
    fileUrl: value.fileUrl,
    imageGroup: ImageGroupToJSON(value.imageGroup),
  }
}